# fix(hub): ensure that an update is dispatched if any of its topics is subscribed and allowed