--- a/subscriber_test.go
+++ b/subscriber_test.go
@@ -10,15 +10,16 @@ import (
 
 func TestDispatch(t *testing.T) {
 	s := NewSubscriber("1", zap.NewNop())
-	s.Topics = []string{"http://example.com"}
+	s.SubscribedTopics = []string{"http://example.com"}
+	s.SubscribedTopics = []string{"http://example.com"}
 	defer s.Disconnect()
 
 	// Dispatch must be non-blocking
 	// Messages coming from the history can be sent after live messages, but must be received first
-	s.Dispatch(&Update{Topics: s.Topics, Event: Event{ID: "3"}}, false)
-	s.Dispatch(&Update{Topics: s.Topics, Event: Event{ID: "1"}}, true)
-	s.Dispatch(&Update{Topics: s.Topics, Event: Event{ID: "4"}}, false)
-	s.Dispatch(&Update{Topics: s.Topics, Event: Event{ID: "2"}}, true)
+	s.Dispatch(&Update{Topics: s.SubscribedTopics, Event: Event{ID: "3"}}, false)
+	s.Dispatch(&Update{Topics: s.SubscribedTopics, Event: Event{ID: "1"}}, true)
+	s.Dispatch(&Update{Topics: s.SubscribedTopics, Event: Event{ID: "4"}}, false)
+	s.Dispatch(&Update{Topics: s.SubscribedTopics, Event: Event{ID: "2"}}, true)
 	s.HistoryDispatched("")
 
 	s.Ready()
@@ -55,4 +56,18 @@ func TestLogSubscriber(t *testing.T) {
 	assert.Contains(t, log, `"remote_addr":"127.0.0.1"`)
 	assert.Contains(t, log, `"topic_selectors":["https://example.com/foo"]`)
 	assert.Contains(t, log, `"topics":["https://example.com/bar"]`)
+}
+
+func TestMatchTopic(t *testing.T) {
+	s := NewSubscriber("", zap.NewNop())
+	s.SetTopics([]string{"https://example.com/no-match", "https://example.com/books/{id}"}, []string{"https://example.com/users/foo/{?topic}"})
+
+	assert.False(t, s.Match(&Update{Topics: []string{"https://example.com/not-subscribed"}}))
+	assert.False(t, s.Match(&Update{Topics: []string{"https://example.com/not-subscribed"}, Private: true}))
+	assert.False(t, s.Match(&Update{Topics: []string{"https://example.com/no-match"}, Private: true}))
+	assert.False(t, s.Match(&Update{Topics: []string{"https://example.com/books/1"}, Private: true}))
+	assert.False(t, s.Match(&Update{Topics: []string{"https://example.com/books/1", "https://example.com/users/bar/?topic=https%3A%2F%2Fexample.com%2Fbooks%2F1"}, Private: true}))
+
+	assert.True(t, s.Match(&Update{Topics: []string{"https://example.com/books/1"}}))
+	assert.True(t, s.Match(&Update{Topics: []string{"https://example.com/books/1", "https://example.com/users/foo/?topic=https%3A%2F%2Fexample.com%2Fbooks%2F1"}, Private: true}))
 }