There is a flaw in the current logic. According to the spec, an update must be dispatched if:

* The subscriber is subscribed to at least one of its topics
* The subscriber is allowed for at least one of its topics (even if it's not the one used to subscribe)

https://mercure.rocks/spec#subscribers

This fixes #651 (thanks @mab05k).

@divine @kevburnsjr, this changes a bit the logic introduced in https://github.com/dunglas/mercure/pull/578.
Would you mean to check it doesn't introduce any performance regression?