@dunglas 2022-08-13
I tweaked the logic again because this patch broke the cache: https://github.com/dunglas/mercure/pull/689