@divine 2022-08-15
Hello,

I've deployed the latest beta version to production. Let's run it for a week.

We're again under server limits... going up to 16M of notifications per day with up to ~4k subscribers. Probably time to upgrade the server 😄.

Thanks!