@divine 2022-08-22
Hello,

JFYI, all looks good. Hard to measure performance (importing thing it doesn't go over server limits) as the number of notifications differs from time to time.

Currently:
![image](https://user-images.githubusercontent.com/48183131/185846457-3f141789-519b-48fd-8dbc-8d7893e7e716.png)

Previously:
![image](https://user-images.githubusercontent.com/48183131/185846395-3258aff6-7efa-4c9f-8b8f-e914414e50c1.png)

Thanks!