@dunglas 2022-08-22
Thanks for testing @divine! I've an idea to improve the algorithm a bit more (using gob instead of ascii86 to encode the key), but this shouldn't change significantly the performance.